api: detect binary execution that raise process capabilities.
Add binary_properties.caps_raised bool if set then the current execution
gained new capabilities through binary execution.

This happens when the executed binary has:
1. setuid root bit set
2. file capabilities.

This is explicitly caps_raised instead of file_caps, as it is hard to
inspect from BPF side if the setuid binary execution is a setuid root
that raised capabilities or not. To identify that we need from BPF side
to check if that uid is explicitly _mapped_ as _root_ id in the current
user namespace or one of the parent user namespaces. Currently this
is not trivial and we still don't do or plan todo user namespace ids
translation in Tetragon.

To solve this we abstract it in `caps_raised` field to indicate if
the binary execution allowed the process to raise its capabilities
which is what users want. The final granted capabilities are listed in
the capabilities field of the process execution.

The new caps_raised is a bool type, no need to rencode the full capabilities
there too, let's just indicate to the user and if they are interested
they can inspect other fields that contain the permitted and effective
capabilities of the current binary.

Signed-off-by: Djalal Harouni <tixxdz@gmail.com>