add tetragon-oci-hook-setup
tetragon-oci-hook-setup is meant to be executed in an init container to
setup run-time hooks on the host. Currently, it only supports the
oci-hooks interface (as, for example, used by cri-o). NRI or other
interfaces fall into future work.

There are two commands: install and uninstall.

Install will copy the hook binary to a directory (which should be a host
mount) and add a configuration file to the hooks directory (also should
be a host mount) to call the binary.

Signed-off-by: Kornilios Kourtis <kornilios@isovalent.com>